Fix: move ArrayGetter into qcodes so it can be used in tests
