+++ qcodes/instrument/mock.py
@@ -3,6 +3,9 @@ import time
 from datetime import datetime
 
 from .base import Instrument
+from .parameter import Parameter
+from qcodes import Loop
+from qcodes.data.data_array import DataArray
 from qcodes.process.server import ServerManager, BaseServer
 from qcodes.utils.nested_attrs import _NoDefault
 
@@ -279,3 +282,36 @@ class MockModel(ServerManager, BaseServer):  # pragma: no cover
         See NestedAttrAccess for details.
         """
         self.ask('method_call', 'delattr', attr)
+
+class ArrayGetter(Parameter):
+    """
+    Example parameter that just returns a single array
+
+    TODO: in theory you can make this same Parameter with
+    name, label & shape (instead of names, labels & shapes) and altered
+    setpoints (not wrapped in an extra tuple) and this mostly works,
+    but when run in a loop it doesn't propagate setpoints to the
+    DataSet. We could track down this bug, but perhaps a better solution
+    would be to only support the simplest and the most complex Parameter
+    forms (ie cases 1 and 5 in the Parameter docstring) and do away with
+    the intermediate forms that make everything more confusing.
+    """
+    def __init__(self, measured_param, sweep_values, delay):
+        name = measured_param.name
+        super().__init__(names=(name,))
+        self._instrument = getattr(measured_param, '_instrument', None)
+        self.measured_param = measured_param
+        self.sweep_values = sweep_values
+        self.delay = delay
+        self.shapes = ((len(sweep_values),),)
+        set_array = DataArray(parameter=sweep_values.parameter,
+                              preset_data=sweep_values)
+        self.setpoints = ((set_array,),)
+        if hasattr(measured_param, 'label'):
+            self.labels = (measured_param.label,)
+
+    def get(self):
+        loop = Loop(self.sweep_values, self.delay).each(self.measured_param)
+        data = loop.run_temp()
+        array = data.arrays[self.measured_param.full_name]
+        return (array,)
